Create forma
fitSharp can process tests written in a plain text format. You may find this simpler to edit and manage than the standard HTML format, depending on your team environment, skills and culture.

We can include plain text format tests on test pages with other information in any format: additional plain text, wiki markup or HTML. The sections of the test page containing our plain text tests begin with "test@" and end with "@test". If the second delimiter is not present, the plain text test continues to the end of the page.

The test is organized into tables, the same as traditional HTML table tests. Each table row is written as one line of text. The table cells are separated by white space. If a cell contains embedded white space, we can wrap the cell content in quotes. The tables are separated by blank lines. Here's a simple test:
test@
23:62:568:2