Update README.md
Improved opening to add in reference to identifiers and a bit more context regarding the history of identity standards. 

I am sorry I didn't mean to commit to the main branch. I thought I was forking and doing it in my branch.. I will try to reverse this and do the thing in my branch and then send a pull request. Github is still a bit mysterious to me I'm doing my best.